v19.0.5.0: Cloud work (#24)
* WIP: First cloud jobs

* Handling max dispacement as float instead of int

* Stage: More fixes

* Cloud: Values for pedal_fixture_z_build.stl

* CloudConnector: Removing a debug print

* CloudProxy: Cleanups and adding material variables

* Stage: Cleanups

* Stage: Main: Making ourselves independent from Bridge.js

* updated material database nus and strengths

* WIP and cleanup

* Adding missing resources

+++ SmartSlicePlugin/SmartSliceExtension/SmartSliceVariables.py
@@ -20,7 +20,7 @@ class SmartSliceVariables(QObject):
 
     maxDeflectChanged = pyqtSignal()
 
-    @pyqtProperty(int, notify=maxDeflectChanged)
+    @pyqtProperty(float, notify=maxDeflectChanged)
     def maxDeflect(self):
         return self._maxDeflect
